// frontend/src/api/generated/models/userDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * acpms-server
 * OpenAPI spec version: 0.1.0
 */

export interface UserDto {
  /** @nullable */
  avatar_url?: string | null;
  created_at: string;
  email: string;
  /** @nullable */
  gitlab_username?: string | null;
  global_roles: string[];
  id: string;
  name: string;
  updated_at: string;
}